backend/cosmetology-app/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.14
# by the following command:
#
#    pip-compile --no-emit-index-url --no-strip-extras requirements.in
#
attrs==25.4.0
    # via
    #   cattrs
    #   jsii
aws-cdk-asset-awscli-v1==2.2.242
    # via aws-cdk-lib
aws-cdk-asset-node-proxy-agent-v6==2.1.0
    # via aws-cdk-lib
aws-cdk-aws-lambda-python-alpha==2.233.0a0
    # via -r requirements.in
aws-cdk-cloud-assembly-schema==48.20.0
    # via aws-cdk-lib
aws-cdk-lib==2.233.0
    # via
    #   -r requirements.in
    #   aws-cdk-aws-lambda-python-alpha
    #   cdk-nag
cattrs==25.3.0
    # via jsii
cdk-nag==2.37.55
    # via -r requirements.in
constructs==10.4.4
    # via
    #   -r requirements.in
    #   aws-cdk-aws-lambda-python-alpha
    #   aws-cdk-lib
    #   cdk-nag
importlib-resources==6.5.2
    # via jsii
jsii==1.125.0
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-node-proxy-agent-v6
    #   aws-cdk-aws-lambda-python-alpha
    #   aws-cdk-cloud-assembly-schema
    #   aws-cdk-lib
    #   cdk-nag
    #   constructs
publication==0.0.3
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-node-proxy-agent-v6
    #   aws-cdk-aws-lambda-python-alpha
    #   aws-cdk-cloud-assembly-schema
    #   aws-cdk-lib
    #   cdk-nag
    #   constructs
    #   jsii
python-dateutil==2.9.0.post0
    # via jsii
pyyaml==6.0.3
    # via -r requirements.in
six==1.17.0
    # via python-dateutil
typeguard==2.13.3
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-node-proxy-agent-v6
    #   aws-cdk-aws-lambda-python-alpha
    #   aws-cdk-cloud-assembly-schema
    #   aws-cdk-lib
    #   cdk-nag
    #   constructs
    #   jsii
typing-extensions==4.15.0
    # via
    #   cattrs
    #   jsii
